requirements.txt
Flask
line-bot-sdk
openai